+++ gdtoolkit/formatter/__main__.py
@@ -1,11 +1,12 @@
 """GDScript formatter
 
 Uncompromising GDScript code formatter. The only configurable thing is
-max line length allowed. The rest will be taken care of by gdformat in a one,
-consistent way.
+max line length allowed and tabs/spaces indent. The rest will be taken
+care of by gdformat in a one, consistent way.
 
 Usage:
   gdformat <path>... [options]
+  gdformat --dump-default-config
 
 Options:
   -c --check                 Don't write the files back,
@@ -16,21 +17,28 @@ Options:
   -f --fast                  Skip safety checks.
   -l --line-length=<int>     How many characters per line to allow.
                              [default: 100]
+  -s --use-spaces=<int>      Use spaces for indent instead of tabs.
   -h --help                  Show this screen.
   --version                  Show version.
+  --dump-default-config      Dump default config to 'gdformatrc' file.
 
 Examples:
   echo 'pass' | gdformat -   # reads from STDIN
 """
 import sys
+import os
+import logging
+import pathlib
 import difflib
-from typing import List, Tuple
+from typing import List, Tuple, Optional
+from types import MappingProxyType
 import pkg_resources
 
 from docopt import docopt
 import lark
+import yaml
 
-from gdtoolkit.formatter import format_code, check_formatting_safety
+from gdtoolkit.formatter import format_code, check_formatting_safety, DEFAULT_CONFIG
 from gdtoolkit.formatter.exceptions import (
     TreeInvariantViolation,
     FormattingStabilityViolation,
@@ -43,6 +51,8 @@ from gdtoolkit.common.exceptions import (
     lark_unexpected_input_to_str,
 )
 
+CONFIG_FILE_NAME = "gdformatrc"
+
 
 def main():
     sys.stdout.reconfigure(encoding="utf-8")
@@ -53,33 +63,108 @@ def main():
         ),
     )
 
+    if arguments["--dump-default-config"]:
+        _dump_default_config()
+
     if arguments["--diff"]:
         arguments["--check"] = True
 
     line_length = int(arguments["--line-length"])
+    spaces_for_indent = (
+        int(arguments["--use-spaces"])
+        if arguments["--use-spaces"] is not None
+        else None
+    )
     safety_checks = not arguments["--fast"]
+
+    config_file_path = _find_config_file()
+    config = _load_config_file_or_default(config_file_path)
+    _log_config_entries(config)
+    _update_config_with_missing_entries_inplace(config)
+
     files: List[str] = find_gd_files_from_paths(
-        arguments["<path>"], excluded_directories=set(".git")
+        arguments["<path>"], excluded_directories=set(config["excluded_directories"])
     )
 
     if files == ["-"]:
-        _format_stdin(line_length, safety_checks)
+        _format_stdin(line_length, spaces_for_indent, safety_checks)
     elif arguments["--check"]:
-        _check_files_formatting(files, line_length, arguments["--diff"], safety_checks)
+        _check_files_formatting(
+            files, line_length, spaces_for_indent, arguments["--diff"], safety_checks
+        )
     else:
-        _format_files(files, line_length, safety_checks)
+        _format_files(files, line_length, spaces_for_indent, safety_checks)
 
 
-def _format_stdin(line_length: int, safety_checks: bool) -> None:
+def _dump_default_config() -> None:
+    # TODO: error handling
+    assert not os.path.isfile(CONFIG_FILE_NAME)
+    with open(CONFIG_FILE_NAME, "w", encoding="utf-8") as handle:
+        handle.write(yaml.dump(DEFAULT_CONFIG.copy()))
+    sys.exit(0)
+
+
+def _find_config_file() -> Optional[str]:
+    search_dir = pathlib.Path(os.getcwd())
+    config_file_path = None
+    while search_dir != pathlib.Path(os.path.abspath(os.sep)):
+        file_path = os.path.join(search_dir, CONFIG_FILE_NAME)
+        if os.path.isfile(file_path):
+            config_file_path = file_path
+            break
+        file_path = os.path.join(search_dir, ".{}".format(CONFIG_FILE_NAME))
+        if os.path.isfile(file_path):
+            config_file_path = file_path
+            break
+        search_dir = search_dir.parent
+    return config_file_path
+
+
+def _load_config_file_or_default(config_file_path: Optional[str]) -> MappingProxyType:
+    # TODO: error handling
+    if config_file_path is not None:
+        logging.info("Config file found: '%s'", config_file_path)
+        with open(config_file_path, "r", encoding="utf-8") as handle:
+            return yaml.load(handle.read(), Loader=yaml.Loader)
+
+    logging.info("""No 'gdformatrc' nor '.gdformatrc' found. Using default config...""")
+    return DEFAULT_CONFIG
+
+
+def _log_config_entries(config: MappingProxyType) -> None:
+    logging.info("Loaded config:")
+    for entry in config.items():
+        logging.info(entry)
+
+
+def _update_config_with_missing_entries_inplace(config: dict) -> None:
+    for key in DEFAULT_CONFIG:
+        if key not in config:
+            logging.info(
+                "Adding missing entry from defaults: %s", (key, DEFAULT_CONFIG[key])
+            )
+            config[key] = DEFAULT_CONFIG[key]
+
+
+def _format_stdin(
+    line_length: int, spaces_for_indent: Optional[int], safety_checks: bool
+) -> None:
     code = sys.stdin.read()
-    success, _, formatted_code = _format_code(code, line_length, "STDIN", safety_checks)
+    success, _, formatted_code = _format_code(
+        code, line_length, spaces_for_indent, "STDIN", safety_checks
+    )
     if not success:
         sys.exit(1)
     print(formatted_code, end="")
 
 
+# pylint: disable-next=too-many-locals
 def _check_files_formatting(
-    files: List[str], line_length: int, print_diff: bool, safety_checks: bool
+    files: List[str],
+    line_length: int,
+    spaces_for_indent: Optional[int],
+    print_diff: bool,
+    safety_checks: bool,
 ) -> None:
     formattable_files = set()
     failed_files = set()
@@ -88,7 +173,7 @@ def _check_files_formatting(
             with open(file_path, "r", encoding="utf-8") as handle:
                 code = handle.read()
                 success, actually_formatted, formatted_code = _format_code(
-                    code, line_length, file_path, safety_checks
+                    code, line_length, spaces_for_indent, file_path, safety_checks
                 )
                 if success and actually_formatted:
                     print(f"would reformat {file_path}", file=sys.stderr)
@@ -135,7 +220,12 @@ def _check_files_formatting(
     sys.exit(1)
 
 
-def _format_files(files: List[str], line_length: int, safety_checks: bool) -> None:
+def _format_files(
+    files: List[str],
+    line_length: int,
+    spaces_for_indent: Optional[int],
+    safety_checks: bool,
+) -> None:
     formatted_files = set()
     failed_files = set()
     for file_path in files:
@@ -143,7 +233,7 @@ def _format_files(files: List[str], line_length: int, safety_checks: bool) -> No
             with open(file_path, "r+", encoding="utf-8") as handle:
                 code = handle.read()
                 success, actually_formatted, formatted_code = _format_code(
-                    code, line_length, file_path, safety_checks
+                    code, line_length, spaces_for_indent, file_path, safety_checks
                 )
                 if success and actually_formatted:
                     print(f"reformatted {file_path}")
@@ -173,7 +263,11 @@ def _format_files(files: List[str], line_length: int, safety_checks: bool) -> No
 
 
 def _format_code(
-    code: str, line_length: int, file_path: str, safety_checks: bool
+    code: str,
+    line_length: int,
+    spaces_for_indent: Optional[int],
+    file_path: str,
+    safety_checks: bool,
 ) -> Tuple[bool, bool, str]:
     success = True
     actually_formatted = False
@@ -185,6 +279,7 @@ def _format_code(
         formatted_code = format_code(
             gdscript_code=code,
             max_line_length=line_length,
+            spaces_for_indent=spaces_for_indent,
             parse_tree=code_parse_tree,
             comment_parse_tree=comment_parse_tree,
         )
@@ -195,6 +290,7 @@ def _format_code(
                     code,
                     formatted_code,
                     max_line_length=line_length,
+                    spaces_for_indent=spaces_for_indent,
                     given_code_parse_tree=code_parse_tree,
                     given_code_comment_parse_tree=comment_parse_tree,
                 )