+++ gdtoolkit/formatter/safety_checks.py
@@ -14,6 +14,7 @@ from .exceptions import (
 )
 
 
+# pylint: disable-next=too-many-public-methods
 class LoosenTreeTransformer(Transformer):
     def par_expr(self, args):
         return args[0] if len(args) > 0 else args
@@ -67,6 +68,13 @@ class LoosenTreeTransformer(Transformer):
     def string(self, args):
         string_token = args[0]
         return expression_to_str(string_token)
+
+    def rstring(self, args):
+        string_token = args[0]
+        return expression_to_str(string_token)
+
+    def par_pattern(self, args):
+        return args[0] if len(args) > 0 else args
 
     def signal_stmt(self, args):
         if len(args) > 1 and len(args[1].children) == 0:
@@ -135,12 +143,14 @@ def check_formatting_stability(
     max_line_length: int,
     parse_tree: Optional[Tree] = None,
     comment_parse_tree: Optional[Tree] = None,
+    spaces_for_indent: Optional[int] = None,
 ) -> None:
     code_formatted_again = format_code(
         formatted_code,
         max_line_length,
         parse_tree=parse_tree,
         comment_parse_tree=comment_parse_tree,
+        spaces_for_indent=spaces_for_indent,
     )
     if formatted_code != code_formatted_again:
         diff = "\n".join(