+++ gdtoolkit/gd2py/__init__.py
@@ -13,7 +13,8 @@ def convert_code(gdscript_code: str) -> str:
         gdscript_code, gather_metadata=True
     )  # TODO: is metadata needed?
     context = Context(
-        indent=0,
+        single_indent_size=1,
+        single_indent_string="\t",
         previously_processed_line_number=-1,
         max_line_length=-1,
         gdscript_code_lines=[],
@@ -51,6 +52,7 @@ def _convert_statement(statement: Tree, context: Context) -> List[str]:
         "func_def": _convert_func_def,
         "enum_stmt": _pass,  # TODO: implement
         "classname_stmt": _pass,
+        "property_body_def": _pass,
         "classname_extends_stmt": _pass,
         "signal_stmt": _pass,
         "const_stmt": lambda s, c: [