+++ tests/formatter/input-output-pairs/simple_atom_expressions.in.gd
@@ -21,3 +21,5 @@ func foo():
 	var xb=$%UniqueNodeName/Xyz
 	var xc=%UniqueNodeName
 	var xd=%UniqueNodeName/Xyz
+	var xe=%"a/b/c"
+	var xf=%UniqueNodeName/%AnotherUniqueNodeName